+++ Source/Sitecore.FakeDb/Data/Engines/DataCommands/RemoveVersionCommand.cs
@@ -1,17 +1,43 @@
 ﻿namespace Sitecore.FakeDb.Data.Engines.DataCommands
 {
-  using System;
+  using System.Linq;
 
-  public class RemoveVersionCommand : Sitecore.Data.Engines.DataCommands.RemoveVersionCommand
+  public class RemoveVersionCommand : Sitecore.Data.Engines.DataCommands.RemoveVersionCommand, IDataEngineCommand
   {
+    private DataEngineCommand innerCommand = DataEngineCommand.NotInitialized;
+
+    public void Initialize(DataEngineCommand command)
+    {
+      this.innerCommand = command;
+    }
+
     protected override Sitecore.Data.Engines.DataCommands.RemoveVersionCommand CreateInstance()
     {
-      return new RemoveVersionCommand();
+      return this.innerCommand.CreateInstance<Sitecore.Data.Engines.DataCommands.RemoveVersionCommand, RemoveVersionCommand>();
     }
 
     protected override bool DoExecute()
     {
-      throw new NotImplementedException();
+      var dbitem = this.innerCommand.DataStorage.GetFakeItem(this.Item.ID);
+      var language = this.Item.Language.Name;
+
+      var removed = false;
+
+      foreach (var field in dbitem.Fields)
+      {
+        if (!field.Values.ContainsKey(language))
+        {
+          continue;
+        }
+
+        var langValues = field.Values[language];
+        var lastVersion = langValues.Last();
+
+        langValues.Remove(lastVersion);
+        removed = true;
+      }
+
+      return removed;
     }
   }
 }