+++ Source/Sitecore.FakeDb/Properties/AssemblyVersionInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 
-[assembly: AssemblyVersion("0.10.4.0")]
-[assembly: AssemblyFileVersion("0.10.4.0")]
-[assembly: AssemblyInformationalVersion("0.10.4")]
+[assembly: AssemblyVersion("0.11.0.0")]
+[assembly: AssemblyFileVersion("0.11.0.0")]
+[assembly: AssemblyInformationalVersion("0.11.0")]